Move everything to root dir so all imports can be absolute
