+++ lib/battery_helper.py
